Add a dictionary specifically for Waffle
We had been using merged.dict, but there are too many words in it. It
was creating cases where the Waffle puzzle was not solvable. We need to
delete words that are not in the Waffle lexicon. Instead of deleting
those from the shared dictionary, delete them from our own, private
dictionary.